requirements.txt
anthropic>=0.3.0
tweepy
python-dotenv
pysqlite3-binary
fastapi
hypercorn
asyncio
openai
python-telegram-bot
aiohttp
sqlite-web
gevent
psycopg2
asyncpg
discord-py